pyproject.toml
[tool.poetry]
name = "eora-task"
version = "0.1.0"
description = ""
authors = ["Ivan Bogatyrev <yaneeee@vk.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.1.1"
beautifulsoup4 = "^4.13.4"
hnswlib = "^0.8.0"
nltk = "^3.9.1"
python-telegram-bot = "^22.3"
openai = "^1.99.9"

[tool.poetry.dev-dependencies]
ruff = "^0.12.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
